fix: select codec path by positional-validity allowlist, not denylist
- Stream field-after-length-ref; reject field-after-tail at parse
- Merge 4 duplicate streaming branches to one per backend file
- Make decode_multibyte_unified the SSOT byte-fold for both paths

+++ tests/forge/expected/codec_length_ref_uint16_le_test.c.h
@@ -55,7 +55,10 @@ static inline int test_vector_codec_length_ref_uint16_le(void) {
                     fprintf(stderr,
                         "FAIL: codec_length_ref_uint16_le test_vector @SCXML L25: "
                         "field `payload` expected empty bytes, got len=%zu\n",
-                        decoded.payload_len);
+                        /* The length companion is `size_t` for a tail field but
+                         * the (possibly narrower) wire length-field for a
+                         * length-ref payload — cast so `%zu` is always valid. */
+                        (size_t)decoded.payload_len);
                     ++failures;
                 }
             }
